+++ projects/rf2-better-ui/src/app/components/event-handler/event-handler.component.ts
@@ -12,11 +12,13 @@ export class EventHandlerComponent implements OnInit, OnDestroy {
 
   private tabHolder: HTMLDivElement;
   private observer: MutationObserver;
+  private injector: any;
 
   constructor(private raceButtonService: RaceButtonService) {
   }
 
   ngOnInit(): void {
+    this.injector = angular.element('ui-view').injector();
     this.raceButtonService.addRaceButton();
     waitForElement('nav em').then((e: Element) => {
       const countDown: HTMLElement = document.createElement('rf-race-countdown-timer');
@@ -29,6 +31,12 @@ export class EventHandlerComponent implements OnInit, OnDestroy {
       this.injectChat();
       this.observeTabChange();
     });
+
+    waitForElement('svg.track-map', 2000).then(() => {
+      this.tabHolder = document.querySelector('div[ng-if=\'eventCtrl.gameState.navigationState\']');
+      this.addCarSelect();
+      this.observeTabChange();
+    });
   }
 
   private observeTabChange(): void {
@@ -39,10 +47,37 @@ export class EventHandlerComponent implements OnInit, OnDestroy {
           this.injectChat();
           break;
         }
+
+        if (m.attributeName === 'class' && this.tabHolder.classList.contains('tab-eventinfo')) {
+          this.addCarSelect();
+          break;
+        }
       }
     });
 
     this.observer.observe(this.tabHolder, {attributes: true});
+  }
+
+  private async addCarSelect(): Promise<void> {
+    const holderDiv: HTMLDivElement = document.querySelector('.left-content-bottom div');
+
+    if (holderDiv.querySelector('button.carSelect')) {
+      // already added
+      return;
+    }
+
+    const carSelect: HTMLButtonElement = document.createElement('button');
+    carSelect.classList.add('secondary', 'carSelect');
+    carSelect.innerHTML = `<span>Car Select</span>`;
+
+    const appSwitchService: any = this.injector.get('appSwitchService');
+
+    carSelect.onclick = () => {
+      sessionStorage.setItem('betterUI.carSelect', 'true');
+      appSwitchService.openAppWithTab('race', 'car');
+    };
+
+    holderDiv.appendChild(carSelect);
   }
 
   private injectChat(): void {